initial work to get alloy to work as required module
