+++ Alloy/cli.js
@@ -3,16 +3,10 @@
  * Copyright (c) 2012 by Appcelerator, Inc. All Rights Reserved.
  * See LICENSE for more information on licensing.
  */
-var program = require('commander'),
-	logger = require('./logger'),
-	os = require('os'),
-	U = require('./utils'),
-	colors = require('colors'),
-	_ = require('lodash'),
-	pkginfo = require('pkginfo')(module, 'version'),
-	path = require('path'),
-	fs = require('fs'),
-	CONST = require('./common/constants');
+const program = require('commander');
+const path = require('path');
+const fs = require('fs');
+const colors = require('colors');
 
 // patch to remove the warning in node >=0.8
 path.existsSync = fs.existsSync || path.existsSync;
@@ -78,65 +72,6 @@ program.command('remove'.blue + ' <source>'.white)
 program.parse(process.argv);
 
 
-// Setup up logging output
-Error.stackTraceLimit = Infinity;
-logger.stripColors = program['colors'] === false;
-logger.logLevel = program['logLevel'] || logger.TRACE;
-if (program.config && program.config.indexOf('logLevel') !== -1) {
-	logger.logLevel = -1;
-}
-
-if (!program.noBanner && program.args[0] !== 'info' && (program.config && program.config.indexOf('noBanner') === -1)) {
-	banner();
-}
-
-if (program.args.length === 0) {
-	var help = program.helpInformation();
-	help = help.replace('Usage: turbo COMMAND [ARGS] [OPTIONS]', 'Usage: ' + 'alloy'.blue + ' COMMAND'.white + ' [ARGS] [OPTIONS]'.grey);
-	help = logger.stripColors ? colors.stripColors(help) : help;
-	console.log(help);
-	process.exit(0);
-}
-
-if (program.platform && !_.includes(CONST.PLATFORM_FOLDERS_ALLOY, program.platform)) {
-	U.die('Invalid platform "' + program.platform + '" specified, must be [' + CONST.PLATFORM_FOLDERS_ALLOY.join(',') + ']');
-}
-
-// Validate the given command
-var command = program.args[0];
-if (!_.includes(getCommands(), command)) {
-	U.die('Unknown command: ' + command.red);
-}
-
-// Launch command with given arguments and options
-(require('./commands/' + command + '/index'))(program.args.slice(1), program);
-
-///////////////////////////////
-////////// FUNCTIONS //////////
-///////////////////////////////
-function banner() {
-	const str = `
-     __                __         \n
-    / /_ __  __ _____ / /_   ____ \n
-   / __// / / // ___// __ \\ / __ \\\n
-  / /_ / /_/ // /   / /_/ // /_/ /\n
-  \\__/ \\__,_//_/   /_.___/ \\____/ \n
-                              `;
-
-	if (!program.dump) {
-		console.log(logger.stripColors ? str : str.blue);
-		var m = 'Titanium Turbo ' + module.exports.version + ' by Axway Developer Community.  A mobile framework for Titanium.\n'.white;
-		console.log(logger.stripColors ? colors.stripColors(m) : m);
-	}
-}
-
-function getCommands() {
-	try {
-		var commandsPath = path.join(__dirname, 'commands');
-		return _.filter(fs.readdirSync(commandsPath), function(file) {
-			return path.existsSync(path.join(commandsPath, file, 'index.js'));
-		});
-	} catch (e) {
-		U.die('Error getting command list', e);
-	}
-}
+const Turbo = require('./Turbo');
+const turbo = new Turbo({program});
+turbo.execute();
