+++ bin/alloy
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 
-require('../Alloy/alloy');
+require('../Alloy/cli');
